+++ ScenarioGUI/gui_classes/gui_structure_classes/text_box.py
@@ -104,21 +104,23 @@ class TextBox(Option):
         """
         return self.get_value() != self.wrong_value
 
-    def check_linked_value(self, value: str) -> bool:
+    def check_linked_value(self, value: str, value_if_hidden: bool | None = None) -> bool:
         """
         This function checks if the linked "option" should be shown.
 
         Parameters
         ----------
         value : str
-            str on which the option should be shown
+            string on which the option should be shown
+        value_if_hidden: bool | None
+            the return value, if the option is hidden
 
         Returns
         -------
         bool
             True if the linked "option" should be shown
         """
-        return self.get_value() == value
+        return self.check_value_if_hidden(self.get_value() == value, value_if_hidden)
 
     def create_function_2_check_linked_value(self, value: str, value_if_hidden: bool | None = None) -> Callable[[], bool]:
         """
@@ -127,7 +129,7 @@ class TextBox(Option):
         Parameters
         ----------
         value : str
-            str on which the option should be shown
+            string on which the option should be shown
         value_if_hidden: bool | None
             the return value, if the option is hidden
 
@@ -135,14 +137,14 @@ class TextBox(Option):
         -------
         function
         """
-        return _create_function_2_check_linked_value(self, value, value_if_hidden)
+        return ft_partial(self.check_linked_value, value, value_if_hidden)
 
     def create_widget(
         self,
         frame: QtW.QFrame,
         layout_parent: QtW.QLayout,
-        row: int = None,
-        column: int = None,
+        row: int | None = None,
+        column: int | None = None,
     ) -> None:
         """
         This functions creates the FloatBox widget in the frame.
@@ -153,10 +155,10 @@ class TextBox(Option):
             The frame object in which the widget should be created
         layout_parent : QtW.QLayout
             The parent layout of the current widget
-        row : int
+        row : int | None
             The index of the row in which the widget should be created
             (only needed when there is a grid layout)
-        column : int
+        column : int | None
             The index of the column in which the widget should be created
             (only needed when there is a grid layout)